requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
torch==2.0.0
transformers==4.31.0
pydantic==1.10.12
aiohttp==3.8.5
numpy==1.24.3
jinja2==3.1.2
pyyaml==6.0.1
python-multipart==0.0.6 